+++ src/sunsynk/__init__.py
@@ -11,7 +11,7 @@ from .sunsynk import Sunsynk
 
 # pylint: enable=unused-import
 
-VERSION = "0.5.3"
+VERSION = "0.5.4"
 
 CELSIUS: Final = "°C"
 KWH: Final = "kWh"